% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_shinyDRAC.R
\name{call_shinyDRAC}
\alias{call_shinyDRAC}
\title{ShinyDRAC application}
\usage{
call_shinyDRAC()
}
\description{
This function calls the 'shinyDRAC' application.
}
\examples{
\dontrun{
call_ShinyDRAC()
}
}
\author{
David Strebler, University of Cologne (Germany).
}
