% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_shinyTLdating.R
\name{call_shinyTLdating}
\alias{call_shinyTLdating}
\title{ShinyTLdating application}
\usage{
call_shinyTLdating()
}
\description{
This function calls the 'shinyTLdating' application.
}
\examples{
\dontrun{
call_shinyTLdating()
}
}
\author{
David Strebler, University of Cologne (Germany).
}
