% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLdating2-package.R
\docType{package}
\name{TLdating-package}
\alias{TLdating-package}
\alias{TLdating2}
\alias{TLdating2-package}
\title{Tools for thermoluminescences dating}
\description{
A series of functions for thermoluminescence dating using the MAAD or the
SAR protocol. This package adds to the R package "luminescence".
}
\details{
\tabular{ll}{  Package: \tab TLdating2 \cr
               Type: \tab Package\cr
               Version: \tab 0.0.1 [upcoming] \cr
               Date: \tab 2015-05-XX \cr
               License: \tab GPL-3
            }
}
\author{
\bold{Authors} \cr
 David Strebler, University of Cologne, Germany

 \bold{Beta-tester} \cr
 Anja Zander, University of Cologne, Germany






 \bold{Related package projects}\cr
   \url{http://www.r-luminescence.de} \cr
   \url{http://cran.r-project.org/web/packages/Luminescence}\cr
   \url{http://cran.r-project.org/web/packages/RLumShiny}\cr

 \bold{Package maintainer}
 David Strebler, Geographisches Institut, Universitat zu Koeln, Cologne, Germany. \cr
 \email{david.strebler@uni-koeln.de}
}
\keyword{package}

